pyproject.toml
[project]
name = "class-asigments-1-9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.7.2",
    "openai-agents>=0.2.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
